Will need to test with more processors to see gains, likely
Right now it looks like a low flatness with as many windows
as possible reduces observable kinks.

+++ src/util/random_number_generators.hpp
@@ -12,12 +12,12 @@
 template<typename data_t>
 struct random_number_generator
 {
-    const std::uint32_t my_seed;
+    const std::uint64_t my_seed;
     std::mt19937 generator;
     std::uniform_real_distribution<data_t> u_real;
     // TODO: Should I add a uniform int distribution too? 
     
-    random_number_generator(const std::uint32_t _seed) : my_seed(_seed), generator(my_seed), u_real(0., 1.) 
+    random_number_generator(const std::uint64_t _seed) : my_seed(_seed), generator(my_seed), u_real(0., 1.) 
     {}
     ~random_number_generator() {}
